Remove PublishablePostEditor and merge it into PostEditor
